CP-261 Contracts table renaming and filtering
- [x] Acceptable list of buyers in filter, when viewing individual Supplier's contracts table
- [x] Acceptable list of suppliers in filter, when viewing individual Buyer's contracts table
