+++ src/components/Tables/index.js
@@ -1,4 +1,4 @@
-export { default as TenderTable } from './TenderTable'
+export { default as ContractTable } from './ContractTable'
 export { default as SupplierTable } from './SupplierTable'
 export { default as BuyerTable } from './BuyerTable'
 export { default as ProductTable } from './ProductTable'
