+++ src/layouts/pages/GlobalOverview/tabs/GlobalContracts.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { RedFlagSummary } from "../../../../components/Visualizations"
-import { TenderTable } from '../../../../components/Tables'
+import { ContractTable } from '../../../../components/Tables'
 
 const GlobalContracts = () => {
     function renderMainVisualization() {
@@ -8,7 +8,7 @@ const GlobalContracts = () => {
     }
 
     function renderTable() {
-        return (<TenderTable />)
+        return (<ContractTable />)
     }
 
     return (
